Support for both Kafka 0.8 and Kafka 0.10
