+++ spark-kafka-0-10-writer/src/test/resources/log4j.properties
